pyproject.toml
[project]
name = "googlecloud-doc-pilot"
version = "0.1.0"
description = "A tool to search and navigate Google Cloud documentation"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.0.0",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
    "slack-bolt>=1.27.0",
]
